+++ packages/article/src/__tests__/ios/article.ios.test.js
@@ -1,5 +1,7 @@
 import shared from "../shared";
+import nativeShared from "../shared-native";
 
-describe("Article test on ios", () => {
+describe("Article tests on ios", () => {
   shared();
+  nativeShared();
 });