+++ packages/styleguide/src/__tests__/ios/styleguide.ios.test.js
@@ -1,5 +1,7 @@
 import shared from "../shared";
+import sharedNative from "../shared-native";
 
 describe("Styleguide tests for ios", () => {
   shared();
+  sharedNative();
 });