+++ packages/brightcove-video/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.13"></a>
+## [2.1.13](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@2.1.10...@times-components/brightcove-video@2.1.13) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/brightcove-video
+
+<a name="2.1.12"></a>
+## [2.1.12](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@2.1.10...@times-components/brightcove-video@2.1.12) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/brightcove-video
+
+<a name="2.1.11"></a>
+## [2.1.11](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@2.1.10...@times-components/brightcove-video@2.1.11) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/brightcove-video
+
+<a name="2.1.10"></a>
+## [2.1.10](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@2.1.9...@times-components/brightcove-video@2.1.10) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/brightcove-video
+
+<a name="2.1.9"></a>
+## [2.1.9](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@2.1.8...@times-components/brightcove-video@2.1.9) (2018-04-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/brightcove-video
+
+<a name="2.1.8"></a>
+## [2.1.8](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@2.1.7...@times-components/brightcove-video@2.1.8) (2018-04-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/brightcove-video
+
 <a name="2.1.7"></a>
 ## [2.1.7](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@2.1.6...@times-components/brightcove-video@2.1.7) (2018-04-06)